체크박스 하긴 했는데, json 파일에 연동이 안 됨

+++ src/model/setting/Setting.java
@@ -27,7 +27,7 @@ public interface Setting {
 
     void setGameMode(int gameMode);
 
-    int getDisplayMode();
+    long getDisplayMode();
 
     void setDisplayMode(int displayMode);
 }
